render queue done 50%

+++ src/Render/Base/IndexBuffer.cpp
@@ -8,7 +8,7 @@
 CFEngineRender::IndexBuffer::IndexBuffer(std::vector<unsigned int>& indices) : GLResource(){
     glGenBuffers(1, &handle_);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, handle_);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(float) * indices.size(), indices.data(), GL_STATIC_DRAW);
+    glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(int) * indices.size(), indices.data(), GL_STATIC_DRAW);
 }
 
 void CFEngineRender::IndexBuffer::release() {
